Refactored graphviz visualization codepath.
Moving this into it's own utility module makes the main module easier to
work on.
