+++ tests/check_system.py
@@ -1,4 +1,4 @@
-from p4pktgen.main import process_json_file
+from p4pktgen.main import generate_test_cases
 from p4pktgen.config import Config
 from p4pktgen.core.translator import TestPathResult
 
@@ -6,7 +6,7 @@ from p4pktgen.core.translator import TestPathResult
 class CheckSystem:
     def check_demo1b(self):
         Config().load_test_defaults()
-        results = process_json_file('examples/demo1b.json')
+        results = generate_test_cases('examples/demo1b.json')
         expected_results = {
             ('start', 'sink', (u'node_2', (True, (u'demo1b.p4', 141, u'hdr.ipv4.isValid()')))):
             TestPathResult.NO_PACKET_FOUND,
@@ -35,7 +35,7 @@ class CheckSystem:
 
     def check_demo1(self):
         Config().load_test_defaults()
-        results = process_json_file(
+        results = generate_test_cases(
             'examples/demo1-action-names-uniquified.p4_16.json')
         expected_results = {
             ('start', 'sink', (u'ingress.ipv4_da_lpm', u'ingress.set_l2ptr')):
@@ -55,7 +55,7 @@ class CheckSystem:
 
     def check_demo1_no_uninit_reads(self):
         Config().load_test_defaults()
-        results = process_json_file(
+        results = generate_test_cases(
             'examples/demo1-no-uninit-reads.p4_16.json')
         expected_results = {
             ('start', 'parse_ipv4', 'sink', (u'tbl_demo1nouninitreads120', u'demo1nouninitreads120'), (u'node_3', (True, (u'demo1-no-uninit-reads.p4_16.p4', 121, u'hdr.ipv4.isValid()'))), (u'ingress.ipv4_da_lpm', u'ingress.my_drop'), (u'node_5', (True, (u'demo1-no-uninit-reads.p4_16.p4', 123, u'!meta.fwd_metadata.dropped')))):
@@ -79,7 +79,7 @@ class CheckSystem:
 
     def check_demo9b(self):
         Config().load_test_defaults()
-        results = process_json_file('examples/demo9b.json')
+        results = generate_test_cases('examples/demo9b.json')
         expected_results = {
             ('start', 'parse_ethernet', 'sink', (u'node_2', (False, (u'demo9b.p4', 157, u'hdr.ipv6.version != 6')))):
             TestPathResult.UNINITIALIZED_READ,
@@ -120,7 +120,7 @@ class CheckSystem:
 
     def check_config_table(self):
         Config().load_test_defaults()
-        results = process_json_file('examples/config-table.json')
+        results = generate_test_cases('examples/config-table.json')
         expected_results = {
             ('start', 'sink', (u'ingress.switch_config_params', u'ingress.set_config_parameters'), (u'ingress.mac_da', u'ingress.set_bd_dmac_intf')):
             TestPathResult.UNINITIALIZED_READ,
@@ -143,7 +143,7 @@ class CheckSystem:
 
     def check_demo1_rm_header(self):
         Config().load_test_defaults()
-        results = process_json_file(
+        results = generate_test_cases(
             'examples/demo1_rm_header.json')
         expected_results = {
             ('start', 'parse_ipv4', 'sink', (u'tbl_demo1_rm_header83', u'demo1_rm_header83')):
@@ -155,7 +155,7 @@ class CheckSystem:
 
     def check_add_remove_header(self):
         Config().load_test_defaults()
-        results = process_json_file(
+        results = generate_test_cases(
             'examples/add-remove-header.json')
         expected_results = {
             ('start', 'parse_ipv4', 'sink', (u'node_2', (True, (u'add-remove-header.p4', 136, u'hdr.ipv4.isValid()'))), (u'ingress.ipv4_da_lpm', u'ingress.set_l2ptr'), (u'node_4', (True, (u'add-remove-header.p4', 138, u'!hdr.outer_ipv4.isValid()'))), (u'ingress.mac_da', u'ingress.set_bd_dmac_intf')):
@@ -187,7 +187,7 @@ class CheckSystem:
         Config().load_test_defaults()
         # This test case exercises variable-length extract, lookahead,
         # and verify statements in the parser.
-        results = process_json_file(
+        results = generate_test_cases(
             'examples/checksum-ipv4-with-options.json')
         expected_results = {
             ('start', u'parse_ipv4', u'parse_tcp', 'sink', (u'node_2', (True, (u'checksum-ipv4-with-options.p4', 125, u'hdr.ipv4.isValid() && hdr.tcp.isValid()'))), (u'node_3', (True, (u'checksum-ipv4-with-options.p4', 130, u'hdr.ipv4.ihl == 14')))):
@@ -215,7 +215,7 @@ class CheckSystem:
         # when taken, make certain paths through ingress impossible.
         # Note that there are no test cases containing the state
         # parse_unreachable_state in the parser paths.
-        results = process_json_file(
+        results = generate_test_cases(
             'examples/parser-impossible-transitions.json')
         expected_results = {
             ('start', 'parse_good', 'sink', (u'node_2', (False, (u'parser-impossible-transitions.p4', 92, u'meta.fwd_metadata.parse_status == 0')))):
@@ -256,7 +256,7 @@ class CheckSystem:
         # Similar to the previous test case, this test case has
         # several parser paths that are impossible to traverse, and
         # several that are possible.
-        results = process_json_file(
+        results = generate_test_cases(
             'examples/parser-impossible-transitions2.json')
         expected_results = {
             ('start', 'sink', (u'node_2', (False, (u'parser-impossible-transitions2.p4', 110, u'hdr.ethernet.isValid()')))):
@@ -311,7 +311,7 @@ class CheckSystem:
         Config().load_test_defaults(solve_for_metadata=True,
                                     run_simple_switch=False)
 
-        results = process_json_file('examples/user-metadata.json')
+        results = generate_test_cases('examples/user-metadata.json')
         expected_results = {
             ('start', 'sink', (u'node_2', (False, (u'user-metadata.p4', 81, u'h.e.soui != 0xf53'))), (u'node_4', (False, (u'user-metadata.p4', 83, u'm.meta_field >> 8 == h.e.soui')))):
             TestPathResult.SUCCESS,
@@ -331,7 +331,7 @@ class CheckSystem:
 
         Config().load_test_defaults()
 
-        results = process_json_file('examples/header-stack-variable-length.json')
+        results = generate_test_cases('examples/header-stack-variable-length.json')
         expected_results = {
             ('start', 'sink', (u'tbl_headerstackvariablelength45', u'headerstackvariablelength45')):
             TestPathResult.SUCCESS,
@@ -344,7 +344,7 @@ class CheckSystem:
 
         Config().load_test_defaults()
 
-        results = process_json_file('examples/parser-cycle.json')
+        results = generate_test_cases('examples/parser-cycle.json')
         expected_results = {
             ('start', 'sink', (u'tbl_parsercycle37', u'parsercycle37')):
             TestPathResult.SUCCESS,
@@ -363,7 +363,7 @@ class CheckSystem:
     # state A to parser state B.
     def xfail_parser_parallel_paths(self):
         Config().load_test_defaults()
-        results = process_json_file('examples/parser-parallel-paths.json')
+        results = generate_test_cases('examples/parser-parallel-paths.json')
         expected_results = {
         }
         assert results == expected_results
